chore: apply prettier to `.solc` contract files
`npx prettier contracts/**/*.sol --write`

+++ contracts/Helpers/TargetContract.sol
@@ -2,17 +2,17 @@
 pragma solidity ^0.8.0;
 
 contract TargetContract {
-
     /* solhint-disable */
-    uint number = 5;
+    uint256 number = 5;
     string name = "Simple Contract Name";
+
     /* solhint-enable */
 
-    function getNumber() public view returns (uint) {
+    function getNumber() public view returns (uint256) {
         return number;
     }
 
-    function setNumber(uint _newNumber) public {
+    function setNumber(uint256 _newNumber) public {
         number = _newNumber;
     }
 
@@ -27,5 +27,4 @@ contract TargetContract {
     function revertCall() public pure {
         revert("TargetContract:revertCall: this function has reverted!");
     }
-    
 }
